Add Rust-based SOS filtering and enhance sync detection logic
- Introduced `rust_utils.py` for Rust-implemented SOS filtering functions, allowing for improved performance in filtering operations.
- Updated `sync.pyx` to enhance field order detection with additional confidence metrics for interlaced and progressive fields.
- Modified data types in `sync.pyx` to use `double` for more precise calculations and adjusted related logic accordingly.
- Refactored `utils.py` to utilize the new Rust filtering functions and improved compatibility with different versions of SciPy.
- Enhanced the `StackableMA` class to use Numba's typed lists for better performance and memory management.

+++ cpp-prototype/include/vhsdecode/tbc_scaler.hpp
@@ -73,7 +73,7 @@ public:
      */
     std::vector<RealArray> scaleField(
         const RealArray& input,
-        const std::vector<size_t>& lineStarts,
+        const std::vector<double>& lineStarts,
         size_t numLines);
     
     /**
